Integrate tox (#10)
Use tox to test the package against multiple Django versions and
multiple Python versions. Include the nightly version of Django.

Update TravisCI config to forego Django 1.9 (no longer supported), but
to test on Django nightly as well.

Travis tests against Python 3.7 (development) as well. The task of
installing Python development is onerous, and so we do not include it in
tox.