Integrate coverage and codecov (#16)
NB: We had previously used `setup.py test` to test on TravisCI, allowing
for tests to be packaged and run in a single go. However, coverage gives
funny results if a file is completely untested (0%), simply omitting the
file, which is undesirable. As such we are now calling runtests.py
directly. However, this means we must also manually install the package.
We have the option to use either `pip install -e .` or else `setup.py
develop`, and are using the latter in this case.