Integrate Django tests into project (#18)
Using the master branch of Django (current release 1.11.3), I've gone
through all of the tests and checks of the contrib auth module. I've
ported over tests that either (1) test behavior of our (very similar)
code or (2) tests that ensure parity of behavior and security when using
this package. The goal is to provide a User that feels like Django's
with only minor tweaks.

The goal was to be overly aggressive. There are likely tests that we
could prune from this commit. However, the goal was to ensure that there
were are no tests in Django that do not exist here that could benefit
this package. Of course, this is detailed oriented and a little fiddly,
and I am only human, so there's bound to be something I missed.

See HISTORY.rst for the list of all bugfixes and behavior changes.