Fix #36 - Refactor model mixins into module (#46)
This change places all of the mix-in classes used to build the User
model in a separate module within the package. Instead of accessing them
in improved_user.models, they are now accessed in improved_user.mixins.
Similarly, the UserManager is moved to improved_user.managers. All other
changes are made to reflect this refactor.

To properly test the refactor, a second example project is created. This
project uses the mixins to create a custom User model without putting
improved_user in INSTALLED_APPS. We provide basic tests to check that
the new User model is created, that migrations apply, and that the
UserManager works.

The original example project is updated along the way, based on code
from the new example project, which is generated using Django 2.0a1.

Finally, documentation is changed to reflect these changes, and a
tutorial is provided to rapidly explain how the new example project can
be mimicked.